third_party/WebKit/LayoutTests/storage/indexeddb/deletedatabase-delayed-by-open-and-versionchange-workers-expected.txt

[Worker] Test that deleteDatabase is delayed if a VERSION_CHANGE transaction is running

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/deletedatabase-delayed-by-open-and-versionchange.js
[Worker] indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;
[Worker] 
[Worker] dbname = "deletedatabase-delayed-by-open-and-versionchange.js"
[Worker] indexedDB.deleteDatabase(dbname)
[Worker] indexedDB.open(dbname)
[Worker] 
[Worker] prepareDatabase():
[Worker] sawVersionChange = false
[Worker] upgradeTransactionComplete = false
[Worker] h = event.target.result
[Worker] indexedDB.deleteDatabase(dbname)
[Worker] 
[Worker] onVersionChange():
PASS [Worker] event.target.version is 1
PASS [Worker] event.oldVersion is 1
PASS [Worker] event.newVersion is null
[Worker] sawVersionChange = true
[Worker] Connection is not closed, so 'blocked' should fire
[Worker] 
[Worker] transactionOnComplete():
[Worker] upgradeTransactionComplete = true
[Worker] 
[Worker] onOpenSuccess():
[Worker] h = event.target.result
[Worker] h.close()
[Worker] Closing too late to prevent the in-flight 'blocked' event
[Worker] 
[Worker] deleteDatabaseOnBlocked():
PASS [Worker] sawVersionChange is true
[Worker] 
[Worker] deleteDatabaseOnSuccess():
PASS [Worker] upgradeTransactionComplete is true
PASS successfullyParsed is true

TEST COMPLETE
